iface_info file refactored. Now it is a impl, IfaceInfo
