+++ src/utils/mod.rs
@@ -1,8 +1,5 @@
 pub mod displays;
 pub use displays::*;
 
-pub mod iface_info;
-pub use iface_info::*;
-
 pub mod dns;
 pub use dns::get_host_name;
